- Removed Wallgenerator from Town (Not Used Anyway) - Moved All Wallgenerator out of src. - Moved it into newe folder called wallgen in legacy. - Moved the notready folder into the wallgen folder incase Shady one day comes back xD!
