--- src/com/palmergames/bukkit/wallgen/Walled.java
@@ -1,18 +1,0 @@
-package com.palmergames.bukkit.wallgen;
-
-import java.util.List;
-
-public interface Walled {
-
-	public Wall getWall();
-
-	public List<WallSection> getWallSections();
-
-	public void setWallSections(List<WallSection> wallSections);
-
-	public boolean hasWallSection(WallSection wallSection);
-
-	public void addWallSection(WallSection wallSection);
-
-	public void removeWallSection(WallSection wallSection);
-}
